+++ sql/core/src/test/scala/org/apache/spark/sql/sources/v2/InsertIntoTests.scala
@@ -220,7 +220,7 @@ private[v2] trait InsertIntoSQLOnlyTests
         }
       } catch {
         case a: AnalysisException if !supportsDynamicOverwrite =>
-          assert(a.getMessage.contains("Table does not support dynamic overwrite"))
+          assert(a.getMessage.contains("does not support dynamic overwrite"))
       }
     }
   }
